:recycle: PopulationGraphのデータフェッチを別コンポーネントに分離
