+++ components/PopulationGraph/PopulationGraphFetching.tsx
@@ -1,0 +1,103 @@
+import { FC } from "react";
+import {
+  LineChart,
+  Line,
+  CartesianGrid,
+  Tooltip,
+  Legend,
+  ResponsiveContainer,
+  XAxis,
+  YAxis,
+} from "recharts";
+import { useCheckboxContext } from "../../helper/context/SelectedPrefsContext";
+import { usePopulationData } from "../../helper/hooks/usePopulationData";
+import { useGetPrefectureData } from "../../helper/hooks/useGetPrefectureData";
+
+type GraphDataType = {
+  [year: string]: {
+    year: number;
+    [prefCode: string]: number | undefined;
+  };
+};
+
+export const PopulationGraphFetching: FC = () => {
+  const { selectedPrefs } = useCheckboxContext();
+  const { prefectureDataLoading, prefectureDataError, prefectureData } =
+    useGetPrefectureData();
+  const { populationDataLoading, populationDataError, populationData } =
+    usePopulationData(selectedPrefs);
+
+  if (prefectureDataLoading || populationDataLoading) return <p>Loading...</p>;
+  if (prefectureDataError || populationDataError)
+    return (
+      <>
+        <p style={{ color: "red" }}>{prefectureDataError}</p>
+        <p style={{ color: "red" }}>{populationDataError}</p>
+      </>
+    );
+
+  const prefectureNames = prefectureData.reduce((accumulator, pref) => {
+    accumulator[pref.prefCode] = pref.prefName;
+    return accumulator;
+  }, {} as { [key: string]: string });
+
+  const generateGraphData = (): {
+    year: number;
+    [prefCode: string]: number | undefined;
+  }[] => {
+    const graphData: GraphDataType = {};
+
+    for (const [prefCode, data] of Object.entries(populationData)) {
+      data.forEach((yearData) => {
+        if (!graphData[yearData.year]) {
+          graphData[yearData.year] = { year: yearData.year };
+        }
+        graphData[yearData.year][prefCode] = yearData.value;
+      });
+    }
+
+    return Object.values(graphData).sort((a, b) => a.year - b.year);
+  };
+
+  const prefectureCodes = Object.keys(populationData);
+  const graphData = generateGraphData();
+
+  return (
+    <>
+      <div>
+        <h2>選択された都道府県:</h2>
+        <ul>
+          {selectedPrefs.map((prefCode) => (
+            <li key={prefCode}>{prefectureNames[prefCode]}</li>
+          ))}
+        </ul>
+      </div>
+      <ResponsiveContainer width="100%" height={400}>
+        <LineChart
+          data={graphData}
+          margin={{ top: 5, right: 30, left: 20, bottom: 5 }}
+        >
+          <CartesianGrid strokeDasharray="3 3" />
+          <XAxis
+            dataKey="year"
+            label={{ value: "年度", position: "insideBottomRight", offset: 0 }}
+          />
+          <YAxis
+            label={{ value: "人口", angle: -90, position: "insideLeft" }}
+          />
+          <Tooltip />
+          <Legend />
+          {prefectureCodes.map((prefCode, index) => (
+            <Line
+              key={prefCode}
+              type="monotone"
+              dataKey={prefCode}
+              name={prefectureNames[prefCode]}
+              stroke={`hsl(${(index * 137.5) % 360}, 70%, 50%)`}
+            />
+          ))}
+        </LineChart>
+      </ResponsiveContainer>
+    </>
+  );
+};
